+++ studio/src/app/components/editor/actions/element/app-actions-element/app-actions-element.tsx
@@ -93,7 +93,7 @@ export class AppActionsElement {
     this.initWindowResize();
   }
 
-  async componentDidUnload() {
+  async disconnectedCallback() {
     this.removeWindowResize();
   }