+++ studio/src/app/pages/core/app-settings/app-settings.tsx
@@ -123,7 +123,7 @@ export class AppHome {
     }
   }
 
-  componentDidUnload() {
+  disconnectedCallback() {
     this.destroyListener();
   }